Rename parallelizer to api_utils
