+++ api_utils/__init__.py
@@ -1,7 +1,7 @@
 ########################################################
-# ------------- dkulib.parallelizer: 0.1.0 ----------------
+# ------------- dkulib.api_utils: 0.1.0 ----------------
 
-# For more information, see https://github.com/dataiku/dss-plugin-dkulib/tree/main/parallelizer
+# For more information, see https://github.com/dataiku/dss-plugin-dkulib/tree/main/api_utils
 # Library version: 0.1.0
 # Last update: 2021-07-16
 # Author: Dataiku (Alex Combessie, Niklas Muennighoff)
